Added to the unit tests.
Registration controller will now respond with an error if activation is attempted with a token that is not an activation type token, or if the account ID from the token cannot be found in the datastore. The test/lib/token tests now include one to fail JWT signing if the account is undefined. The test/routes/activate tests now include one to fail if the access token contains an invalid account ID.