Tokenwall changes.
Instead of just decoded the token in the tokenwall middleware, now it is also decrypted, and added to the req object as req.accountId. The tokenwall controller now inserts the actual account ID into the fun message, instead of using DZ-015. The tokenwall unit test is adjusted to check for the new message format.