More code coverage
Moved express error handling out to lib/error.js and added to the tokenwalltest route so that it also has tokenwalltest/throws which throws an exception, to test the status 500 error handling middleware. Changed how the app starts, now it starts the database connection, and doesn't continue further until that is connected. Removed a couple try/catch that would never be catching an error.