Added try/catch in app.js main()
Now when it catches an error during the main startup function it will use a new errors.handleFatal() function. Added some test stuff to cover that, and moved things around a bit. Now datastore.js only returns the mongoose.connect function if its not already connected. Changed the fun.graffiti slightly. Updated packages.json versions, but not the should library because it was causing errors, so I left it back on version 11.2.0, and also changed the npm start script to call the main() function specifically.