Some small Mocha changes
Moved the Mocha configuration from mocha.opts into .mocharc.js. I chose to use the .mocharc.js instead of the .yml or .json options, because the Mocha documentation states that the .mocharc.js is the first priority if no custom path was provided for the configuration file. Also, the JavaScript configuration file allows conditional logic which may come in handy.